+++ test/transform/resource/after-ecj/WithInnerAnnotation.java
@@ -1,0 +1,8 @@
+class WithInnerAnnotation {
+  @interface Inner {
+    int bar() default 42;
+  }
+  WithInnerAnnotation() {
+    super();
+  }
+}